Use legacycrypt instead of crypt on Python >= 3.13
