+++ azurelinuxagent/common/utils/textutil.py
@@ -17,12 +17,8 @@
 # Requires Python 2.6+ and Openssl 1.0+
 
 import base64
-# W4901: Deprecated module 'crypt' (deprecated-module)
-import crypt  # pylint: disable=deprecated-module
 import hashlib
-import random
 import re
-import string
 import struct
 import sys
 import traceback
@@ -286,16 +282,6 @@ def remove_bom(c):
        str_to_ord(c[2]) > 128:
         c = c[3:]
     return c
-
-
-def gen_password_hash(password, crypt_id, salt_len):
-    collection = string.ascii_letters + string.digits
-    salt = ''.join(random.choice(collection) for _ in range(salt_len))
-    salt = "${0}${1}".format(crypt_id, salt)
-    if sys.version_info[0] == 2:
-        # if python 2.*, encode to type 'str' to prevent Unicode Encode Error from crypt.crypt
-        password = password.encode('utf-8')
-    return crypt.crypt(password, salt)
 
 
 def get_bytes_from_pem(pem_str):
