+++ requirements.txt
@@ -1,2 +1,3 @@
 distro; python_version >= '3.8'
 pyasn1
+legacycrypt; python_version >= '3.13'
