+++ setup.py
@@ -314,13 +314,16 @@ class install(_install):  # pylint: disable=C0103
 
 
 # Note to packagers and users from source.
-# In version 3.5 of Python distribution information handling in the platform
-# module was deprecated. Depending on the Linux distribution the
-# implementation may be broken prior to Python 3.7 wher the functionality
-# will be removed from Python 3
-requires = []  # pylint: disable=invalid-name
-if sys.version_info[0] >= 3 and sys.version_info[1] >= 7:
-    requires = ['distro']  # pylint: disable=invalid-name
+# * In version 3.5 of Python distribution information handling in the platform
+#   module was deprecated. Depending on the Linux distribution the
+#   implementation may be broken prior to Python 3.8 where the functionality
+#   will be removed from Python 3.
+# * In version 3.13 of Python, the crypt module was removed and legacycrypt is
+#   required instead.
+requires = [
+    "distro;python_version>='3.8'",
+    "legacycrypt;python_version>='3.13'",
+]
 
 modules = []  # pylint: disable=invalid-name
