+++ tests/common/utils/test_text_util.py
@@ -16,7 +16,6 @@
 #
 
 import hashlib
-import os
 import unittest
 from azurelinuxagent.common.future import LooseVersion as Version
 
@@ -26,15 +25,6 @@ from tests.lib.tools import AgentTestCase
 
 
 class TestTextUtil(AgentTestCase):
-    def test_get_password_hash(self):
-        with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'test_passwords.txt'), 'rb') as in_file:
-            for data in in_file:
-                # Remove bom on bytes data before it is converted into string.
-                data = textutil.remove_bom(data)
-                data = ustr(data, encoding='utf-8')
-                password_hash = textutil.gen_password_hash(data, 6, 10)
-                self.assertNotEqual(None, password_hash)
-
     def test_replace_non_ascii(self):
         data = ustr(b'\xef\xbb\xbfhehe', encoding='utf-8')
         self.assertEqual('hehe', textutil.replace_non_ascii(data))
